convert nl80211 commands to string
Tinkering with a way to convert the nl80211 command enum to a string.
Added a way to convert the the parsed enums to a C function. Still need
to generate a header file. Had to manually remove the duplicates;
probably something that I can do in penums.py but I need to finisht he
value parsing first.